requirements.txt
fastapi
uvicorn
sqlalchemy
openai
streamlit
python-dotenv
requests
google-generativeai
